+++ jaorm-processor/src/main/java/io/github/ulisse1996/jaorm/processor/strategy/ParametersStrategy.java
@@ -2,6 +2,7 @@ package io.github.ulisse1996.jaorm.processor.strategy;
 
 import javax.annotation.processing.ProcessingEnvironment;
 import javax.lang.model.element.ExecutableElement;
+import java.util.List;
 import java.util.Set;
 
 public interface ParametersStrategy {